Split hello into main and hello pkg
