+++ cmd/hello/main.go
@@ -1,0 +1,11 @@
+package main
+
+import (
+	"fmt"
+
+	"github.com/eccles/hestia/pkg/hello"
+)
+
+func main() {
+	fmt.Printf("%s", hello.Hello())
+}
