+++ example-commit
@@ -6,5 +6,5 @@ Clear succinct statement of problem.
 Solution:
 Summarised description of solution.
     
-Signed-off-by: User Name <user@email.com>
+Signed-off-by: Paul Hewlett <phewlett76@gmail.com>
