+++ pkg/hello/hello.go
@@ -1,15 +1,9 @@
-package main
+package hello
 
 import (
-	"fmt"
-
 	"rsc.io/quote"
 )
 
 func Hello() string {
 	return quote.Go()
-}
-
-func main() {
-	fmt.Printf("%s", Hello())
 }
