+++ pkg/hello/hello_test.go
@@ -1,4 +1,4 @@
-package main
+package hello
 
 import (
 	"testing"
